refactor(telegram): extract rich message delivery into TelegramRichMixin (adapter god-file slice)

# Conflicts:
#	plugins/platforms/telegram/adapter.py
